requirements.txt
python-dotenv~=0.19.0
SQLAlchemy~=1.4.23
vk_api~=11.9.4
psycopg2~=2.9.1